pyproject.toml
[project]
name = "visualapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioodbc>=0.5.0",
    "click>=8.2.1",
    "fastapi-cache2>=0.2.2",
    "fastapi[standard]>=0.115.12",
    "pydantic>=2.11.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
]

